my_agent/requirements.txt
google-generativeai
langchain-google-genai
langchain-community
pydantic
chromadb
langchain-chroma
langchain-core
pandas
numpy
matplotlib
tiktoken
pypdf 
pymupdf
unstructured
markdown
deepeval
ragas
python-dotenv
langchain_cohere
cohere
langchain_tavily
langchain_unstructured
supabase
upstash-redis
langchain-community 
redis
langchain-openai
langchain-mongodb